Split profile page
